feat: Introduce `noosphere-gateway` crate
