+++ release-please-config.json
@@ -12,6 +12,7 @@
     "rust/noosphere-collections": {},
     "rust/noosphere-core": {},
     "rust/noosphere-fs": {},
+    "rust/noosphere-gateway": {},
     "rust/noosphere-ipfs": {},
     "rust/noosphere-into": {},
     "rust/noosphere-ns": {},
