+++ rust/noosphere-core/src/lib.rs
@@ -1,6 +1,8 @@
 #[macro_use]
-extern crate log;
+extern crate tracing as extern_tracing;
 
 pub mod authority;
 pub mod data;
 pub mod view;
+
+pub mod tracing;
