+++ rust/noosphere-gateway/src/route/did.rs
