+++ rust/noosphere-ipfs/src/kubo.rs
@@ -107,6 +107,7 @@ mod tests {
     use cid::Cid;
     use iroh_car::{CarHeader, CarWriter};
     use libipld_cbor::DagCborCodec;
+    use noosphere_core::tracing::initialize_tracing;
     use noosphere_storage::block_serialize;
     use serde::{Deserialize, Serialize};
     use url::Url;
