+++ rust/noosphere/src/wasm/mod.rs
@@ -2,9 +2,7 @@ mod file;
 mod fs;
 mod noosphere;
 mod sphere;
-mod tracing;
 
-pub use self::tracing::*;
 pub use file::*;
 pub use fs::*;
 pub use noosphere::*;
